Switched to doubles for boundaries and added centralise call

+++ src/main/kotlin/com/manimdsl/linearrepresentation/Objects.kt
@@ -124,10 +124,10 @@ data class VariableBlock(
 sealed class DataStructureMObject(
     open val type: DataStructureType,
     open val ident: String,
-    private var boundaries: List<Pair<Int, Int>> = emptyList()
+    private var boundaries: List<Pair<Double, Double>> = emptyList()
 ) : MObject {
 
-    abstract fun setNewBoundary(corners: List<Pair<Int, Int>>, newMaxSize: Int)
+    abstract fun setNewBoundary(corners: List<Pair<Double, Double>>, newMaxSize: Int)
 
 }
 
@@ -140,7 +140,7 @@ data class InitManimStack(
     val moveToShape: Shape? = null,
     val color: String? = null,
     val textColor: String? = null,
-    private var boundary: List<Pair<Int, Int>> = emptyList(),
+    private var boundary: List<Pair<Double, Double>> = emptyList(),
     private var maxSize: Int = -1
 ) : DataStructureMObject(type, ident, boundary) {
     override var shape: Shape = NullShape
@@ -152,7 +152,7 @@ data class InitManimStack(
         return python
     }
 
-    override fun setNewBoundary(corners: List<Pair<Int, Int>>, newMaxSize: Int) {
+    override fun setNewBoundary(corners: List<Pair<Double, Double>>, newMaxSize: Int) {
         maxSize = newMaxSize
         boundary = corners
         shape = InitManimStackShape(ident, text, boundary, alignment, color, textColor)
@@ -167,7 +167,7 @@ data class ArrayStructure(
     val color: String? = null,
     val textColor: String? = null,
     var maxSize: Int = -1,
-    private var boundaries: List<Pair<Int, Int>> = emptyList()
+    private var boundaries: List<Pair<Double, Double>> = emptyList()
 ) : DataStructureMObject(type, ident, boundaries) {
     override var shape: Shape = NullShape
 
@@ -180,7 +180,7 @@ data class ArrayStructure(
         )
     }
 
-    override fun setNewBoundary(corners: List<Pair<Int, Int>>, newMaxSize: Int) {
+    override fun setNewBoundary(corners: List<Pair<Double, Double>>, newMaxSize: Int) {
         maxSize = newMaxSize
         boundaries = corners
         shape = ArrayShape(ident, values, text, boundaries, color, textColor)
